refactor(kitchenpos): modal 패키지 영역을 domain 패키지 영역으로 재정의
