+++ src/main/java/kitchenpos/menu/domain/MenuDao.java
@@ -1,6 +1,4 @@
-package kitchenpos.menu.model;
-
-import kitchenpos.menu.model.Menu;
+package kitchenpos.menu.domain;
 
 import java.util.List;
 import java.util.Optional;
