+++ src/main/java/kitchenpos/menu/infrastructure/JdbcTemplateMenuDao.java
@@ -1,7 +1,7 @@
 package kitchenpos.menu.infrastructure;
 
-import kitchenpos.menu.model.Menu;
-import kitchenpos.menu.model.MenuDao;
+import kitchenpos.menu.domain.Menu;
+import kitchenpos.menu.domain.MenuDao;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
