+++ src/main/java/kitchenpos/menuGroup/application/MenuGroupBo.java
@@ -1,7 +1,7 @@
 package kitchenpos.menuGroup.application;
 
-import kitchenpos.menuGroup.model.MenuGroupDao;
-import kitchenpos.menuGroup.model.MenuGroup;
+import kitchenpos.menuGroup.domain.MenuGroupDao;
+import kitchenpos.menuGroup.domain.MenuGroup;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
