+++ src/main/java/kitchenpos/menuGroup/web/MenuGroupRestController.java
@@ -1,6 +1,6 @@
 package kitchenpos.menuGroup.web;
 
-import kitchenpos.menuGroup.model.MenuGroup;
+import kitchenpos.menuGroup.domain.MenuGroup;
 import kitchenpos.menuGroup.application.MenuGroupBo;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
