+++ src/main/java/kitchenpos/order/web/OrderRestController.java
@@ -1,6 +1,6 @@
 package kitchenpos.order.web;
 
-import kitchenpos.order.model.Order;
+import kitchenpos.order.domain.Order;
 import kitchenpos.order.application.OrderBo;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
