+++ src/main/java/kitchenpos/product/application/ProductBo.java
@@ -1,7 +1,7 @@
 package kitchenpos.product.application;
 
-import kitchenpos.product.model.ProductDao;
-import kitchenpos.product.model.Product;
+import kitchenpos.product.domain.ProductDao;
+import kitchenpos.product.domain.Product;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
