+++ src/main/java/kitchenpos/product/infrastructure/JdbcTemplateProductDao.java
@@ -1,7 +1,7 @@
 package kitchenpos.product.infrastructure;
 
-import kitchenpos.product.model.Product;
-import kitchenpos.product.model.ProductDao;
+import kitchenpos.product.domain.Product;
+import kitchenpos.product.domain.ProductDao;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
