+++ src/test/java/kitchenpos/bo/InMemoryMenuProductDao.java
@@ -1,7 +1,7 @@
 package kitchenpos.bo;
 
-import kitchenpos.menu.model.MenuProductDao;
-import kitchenpos.menu.model.MenuProduct;
+import kitchenpos.menu.domain.MenuProductDao;
+import kitchenpos.menu.domain.MenuProduct;
 
 import java.util.*;
 import java.util.stream.Collectors;
