+++ src/test/java/kitchenpos/bo/MenuBoTest.java
@@ -1,11 +1,11 @@
 package kitchenpos.bo;
 
 import kitchenpos.menu.application.MenuBo;
-import kitchenpos.menu.model.MenuDao;
-import kitchenpos.menuGroup.model.MenuGroupDao;
-import kitchenpos.menu.model.MenuProductDao;
-import kitchenpos.product.model.ProductDao;
-import kitchenpos.menu.model.Menu;
+import kitchenpos.menu.domain.MenuDao;
+import kitchenpos.menuGroup.domain.MenuGroupDao;
+import kitchenpos.menu.domain.MenuProductDao;
+import kitchenpos.product.domain.ProductDao;
+import kitchenpos.menu.domain.Menu;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
