+++ src/test/java/kitchenpos/bo/MenuGroupBoTest.java
@@ -1,8 +1,8 @@
 package kitchenpos.bo;
 
 import kitchenpos.menuGroup.application.MenuGroupBo;
-import kitchenpos.menuGroup.model.MenuGroupDao;
-import kitchenpos.menuGroup.model.MenuGroup;
+import kitchenpos.menuGroup.domain.MenuGroupDao;
+import kitchenpos.menuGroup.domain.MenuGroup;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
