+++ src/test/java/kitchenpos/bo/OrderBoTest.java
@@ -1,12 +1,12 @@
 package kitchenpos.bo;
 
-import kitchenpos.menu.model.MenuDao;
+import kitchenpos.menu.domain.MenuDao;
 import kitchenpos.order.application.OrderBo;
-import kitchenpos.table.model.OrderTableDao;
-import kitchenpos.order.model.OrderDao;
-import kitchenpos.order.model.OrderLineItemDao;
-import kitchenpos.order.model.Order;
-import kitchenpos.order.model.OrderStatus;
+import kitchenpos.table.domain.OrderTableDao;
+import kitchenpos.order.domain.OrderDao;
+import kitchenpos.order.domain.OrderLineItemDao;
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.OrderStatus;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
