+++ src/test/java/kitchenpos/bo/ProductBoTest.java
@@ -1,8 +1,8 @@
 package kitchenpos.bo;
 
 import kitchenpos.product.application.ProductBo;
-import kitchenpos.product.model.ProductDao;
-import kitchenpos.product.model.Product;
+import kitchenpos.product.domain.ProductDao;
+import kitchenpos.product.domain.Product;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
