+++ src/test/java/kitchenpos/controller/ProductRestControllerTest.java
@@ -1,7 +1,7 @@
 package kitchenpos.controller;
 
 import kitchenpos.product.application.ProductBo;
-import kitchenpos.product.model.Product;
+import kitchenpos.product.domain.Product;
 import kitchenpos.product.web.ProductRestController;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
